Factor out pagination reducer and components
