+++ app/frontend/pagination/infinite_pagination.component.js
