+++ app/frontend/pagination/pagination.component.js
