+++ app/frontend/proposals/proposals.component.js
@@ -3,7 +3,7 @@ import { bindActionCreators }                  from 'redux';
 import { connect }                             from 'react-redux';
 
 import Loading                                 from '../application/loading.component';
-import InfinitePagination                      from '../application/infinite_pagination.component';
+import InfinitePagination                      from '../pagination/infinite_pagination.component';
 import ProposalsHeader                         from './proposals_header.component';
 import ProposalsFilterTabs                     from './proposals_filter_tabs.component';
 import ProposalsSidebar                        from './proposals_sidebar.component';
