+++ app/frontend/proposals/proposals_app.component.js
@@ -22,35 +22,14 @@ import {
   APPEND_PROPOSALS_PAGE
 }                        from './proposals.actions';
 import { proposals }     from './proposals.reducers';
+
 import districts         from '../districts/districts.reducers';
 import categories        from '../categories/categories.reducers';
 import filters           from '../filters/filters.reducers';
 import order             from '../order/order.reducers';
+import pagination        from '../pagination/pagination.reducers';
 
 const createStoreWithMiddleware = applyMiddleware(...middlewares)(createStore);
-
-const pagination = function (state = {}, action) {
-  switch (action.type) {
-    case FETCH_PROPOSALS:
-    case APPEND_PROPOSALS_PAGE:
-      let { 
-        current_page, 
-        next_page, 
-        prev_page, 
-        total_pages, 
-        total_count 
-      } = action.payload.data.meta;
-
-      return {
-        current_page,
-        next_page,
-        prev_page,
-        total_pages,
-        total_count
-      };
-  }
-  return state;
-}
 
 const seed = function (state = getInitialSeedState(), action) {
   switch (action.type) {
